requirements.txt
beautifulsoup4==4.13.4
matplotlib==3.10.3
pandas==2.2.3
pandas_datareader==0.10.0
pytest==7.4.4
pytest.egg==info
PyYAML==6.0.2
PyYAML==6.0.1
PyYAML==6.0.2
seaborn==0.13.2
SQLAlchemy==2.0.34
streamlit==1.37.1
yfinance==0.2.61
typing
